core/elevenlabs_client.py
import os
from elevenlabs.client import ElevenLabs
from dotenv import load_dotenv

load_dotenv()

client = ElevenLabs(
    api_key=os.getenv("ELEVENLABS_API_KEY"),
)

def generate_audio(texto: str, voice_id: str) -> bytes:
    audio = client.text_to_speech.convert(
        voice_id=voice_id,
        model_id="eleven_multilingual_v2",
        text=texto,
        output_format="mp3_44100"
    )
    return b"".join(audio)


