requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-dotenv
langdetect
supabase
python-multipart  # For file uploads
pydantic
